Added the intructor's solution to lessHandler, and kept my own in moreHandler. I also allowed those methods to edit 'price'
The one difference was that he passed in the type parameter in BuildControls, assigning an arrow function to the prop 'lessHandler', and using just the function reference in BuildControl. I had instead passed both the function reference and the type down through to BuildControl, and made the arrow function there - a slightly more inneficient method.